requirements.txt
psycopg2
flask
paho.mqtt